requirements.txt
flask
pillow
requests
beautifulsoup4
python-whois
gunicorn